% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpandGrid}
\alias{ExpandGrid}
\title{Generate a Grid of All Combinations of Input Sequence}
\arguments{
\item{Dim}{Integer. The number of dimensions.}

\item{s}{A numeric vector representing the input sequence.}
}
\value{
A numeric matrix where each row is a combination of the input values.
}
\description{
Generates a grid by expanding an input sequence across dimensions.
}
\examples{
s <- c(1, 2, 3)
grid <- ExpandGrid(2, s)
print(grid)
}
