% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kroneckerProduct}
\alias{kroneckerProduct}
\title{Compute Kronecker Product of Sparse Matrices}
\arguments{
\item{A}{A sparse matrix.}

\item{B}{A sparse matrix.}
}
\value{
A sparse matrix representing the Kronecker product of A and B.
}
\description{
Computes the Kronecker product of two sparse matrices.
}
\examples{
library(Matrix)
A <- Diagonal(4)
B <- Diagonal(4)
result <- kroneckerProduct(A, B)
}
