% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mydstn}
\alias{mydstn}
\title{Discrete Sine Transform (DST-II)}
\arguments{
\item{input}{A numeric vector of type \code{Eigen::VectorXd}.}

\item{Ns}{Integer. The size of the input vector.}
}
\value{
A numeric vector representing the transformed values.
}
\description{
Computes the Discrete Sine Transform (DST-II) of a numeric vector.
}
\examples{
input <- c(1, 2, 3, 4)
result <- mydstn(input, length(input))
}
