\name{YauYauFilter-package}
\alias{YauYauFilter-package}
\alias{YauYauFilter}
\docType{package}
\title{
\packageTitle{YauYauFilter}
}
\description{
\packageDescription{YauYauFilter}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{YauYauFilter}

Maintainer: \packageMaintainer{YauYauFilter}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = YauYauFilter)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{YauYauFilter}}
%\section{Documentation Index}{\packageIndices{YauYauFilter}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
